requirements.txt
# ETF RAG 챗봇 필수 라이브러리

# 데이터 처리
pandas>=1.5.0
numpy>=1.21.0

# 웹 앱
streamlit>=1.28.0

# 시각화
plotly>=5.15.0

# API 클라이언트
requests>=2.28.0

# 환경변수 관리
python-dotenv>=1.0.0

# LLM 연동 (CLOVA)
langchain>=0.1.0
langchain_community>=0.0.10

# 동시성 처리 (캐시 생성용)
# concurrent.futures는 Python 표준 라이브러리이므로 별도 설치 불필요